exame2025/20171165trigger-exame/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-storage-blob
azure-iot-hub
